+++ boot/core/src/boot/task/built_in.clj
@@ -775,9 +775,10 @@
 
 (defn- sift-poms
   [fileset project]
-  (let [prj-match (when project
-                    (let [[gid aid] (util/extract-ids project)]
-                      (str "/" gid "/" aid "/pom.xml")))
+  (let [poms      (->> (core/output-files fileset)
+                       (core/by-name ["pom.xml"]))
+        prj-match (when project
+                    (str "/" (pod/full-id project) "/pom.xml"))
         project?  #(if-not prj-match
                      (:project %)
                      (.endsWith (core/tmp-path %) prj-match))]